fix(test): extract shared SDK mocks to prevent CI-only failures
Both claude-sdk.test.ts and claude-sdk-resume.test.ts registered
mock.module() on the same paths (executor-registry, agent-registry,
agent-directory, claude-agent-sdk). Bun's ESM module cache keeps the
FIRST-loaded mock.module binding for cached modules. On CI, bun
consistently loaded claude-sdk.test.ts first, so its mock instances
were bound to claude-sdk.js. When claude-sdk-resume.test.ts ran later,
its own mock instances never received calls — 7 tests failed.

Locally this didn't manifest because bun's dev module loader behaved
differently (possibly re-binding or different file ordering).

Fix: extract all shared mock instances and mock.module registrations
into _sdk-mocks.ts. Both test files import from this single source,
ensuring they reference the SAME mock objects regardless of load order.
Added resetQueryMock() helper to restore queryMock's default
implementation after claude-sdk.test.ts overrides it.

Tests: 2163 pass, 0 fail (was 7 fail on CI).
